Remove Airtable dependency completely (#1415)
* Misc annotations

* Remove pyairtable requirement

* Remove use of Airtable Interface

* Refactor JsonModelsInterface

* Remove AirtableCard since we will always read from S3 JSON

* Code clean up; and remove AirtableInterface

* More code clean up in the card module

* Remove unused script

* Remove unused constants

* Move code to action scripts

* WIP

* WIP single airtable ops script for actions

* WIP: Refactor insert step

* Remove redundant scripts

* Update workflow to include updating S3 on top of Airtable as well
