--- .github/scripts/update_readme_from_airtable.py
@@ -1,8 +1,0 @@
-import sys
-from ersilia.publish.metadata import ReadmeUpdater
-
-repo_name = sys.argv[1]
-repo_path = sys.argv[2]
-
-rm = ReadmeUpdater(model_id=repo_name, repo_path=repo_path, commit=False)
-rm.update()
