+++ ersilia/utils/exceptions_utils/hubdata_exceptions.py
@@ -1,6 +1,6 @@
 from .exceptions import ErsiliaError
 
-
+# Note: Not really used anywhere right now except in the sanitize class
 class InvalidUrlInAirtableError(ErsiliaError):
     def __init__(self, url):
         self.message = "There is an invalid URL in AirTable: {0}".format(url)
